refactor: isolate cache identity
