+++ docs/superpowers/plans/2026-08-17-modularization-plan.md
@@ -65,11 +65,11 @@
 
 **Files:** `lib/MeshCache.lua`, new `lib/CacheStorage.lua`, new `lib/CacheIdentity.lua`, new `lib/CacheManifest.lua`, cache tests
 
-- [ ] Characterize cache file names, metadata, compression, validation, fingerprints, and incomplete-build resume records.
+- [x] Characterize cache file names, metadata, compression, validation, fingerprints, and incomplete-build resume records.
 - [ ] Move filesystem/codec operations into `CacheStorage`.
-- [ ] Move build identity and invalidation rules into `CacheIdentity`.
+- [x] Move build identity and invalidation rules into `CacheIdentity`.
 - [ ] Move manifest and resume-record handling into `CacheManifest`.
-- [ ] Keep `MeshCache` as the stable façade and compare serialized outputs against the baseline fixtures.
+- [x] Keep `MeshCache` as the stable façade and compare serialized outputs against the baseline fixtures.
 
 ## Task 6: Split geometry, queue policy, and GPU ownership behind `ChunkMesher`
